+++ tests/error_level.rs
@@ -4,7 +4,11 @@ extern crate stderrlog;
 
 #[test]
 fn error_level() {
-    stderrlog::new().module(module_path!()).verbosity(0).init().unwrap();
+    stderrlog::new()
+        .module(module_path!())
+        .verbosity(0)
+        .init()
+        .unwrap();
 
     error!("error msg");
     warn!("warning msg");
@@ -12,5 +16,5 @@ fn error_level() {
     debug!("debug msg");
     trace!("trace msg");
 
-    assert_eq!(log::LogLevel::Error, log::max_log_level())
+    assert_eq!(log::Level::Error, log::max_level())
 }